Programma/requirements.txt
click==7.1.2
colorama==0.4.4
dnspython==1.16.0
eventlet==0.30.0
greenlet==0.4.17
shellingham==1.3.2
six==1.15.0
sparql-client==3.8
typer==0.3.2
